// src/gql/__generated__/categories.ts
export type ProductFragment = {
  id: string;
  slug: string;
  name: string;
  description: string | null;
  price: number | null;
  mainAsset: {
    url: string;
  };
  assets: {
    url: string;
  }[];
  category: {
    name: string;
    slug: string;
  } | null;
  isNew: boolean;
  colors: ColorsFragment[];
  sizes: SizesFragment[];
  positionnings: PositionningsFragment[];
  frameColors: FrameColorsFragment[];
  soundbarColors: SoundbarColorsFragment[];
  supportColors: SupportColorsFragment[];
  quantity: number;
  shops: ShopFragment[];
  assetsByProductVariants: AssetsByProductVariantFragment[];
};
export type CategoryFragment = {
  id: string;
  name: string;
  slug: string;
  file: {
    url: string;
  };
  products: ProductFragment[];
};
export type Categories = {
  categories: CategoryFragment[];
};